Refactor: auth.ts client를 supabaseClientClient로 변경
- Home.tsx의 session 불러오는 server action 함수 사용
