+++ lib/admin.ts
@@ -1,10 +1,11 @@
-import { supabase } from './auth';
+import { supabaseClientClient } from './auth';
 
 export type Themes = 'light' | 'dark' | 'system';
 
 export const changeRawTheme = async (theme: Themes, admin_id: string) => {
-  const { data, error } = await supabase
+  const { data, error } = await supabaseClientClient
     .from('theme')
     .insert({ theme, admin_id });
-  console.log(data, error);
+
+  if (error !== null) throw new Error('change theme is error');
 };
